arm: basic inmates demos
This patch adds the necessary libraries for writing simple inmates on arm,
and provides two demos.
It attempts to use the same layout as x86, and allows to set the devices
base addresses with platform-specific includes, in order to avoid
including kconfig.h

Only the vexpress platform with GICv2 and GICv3 is supported for the
moment.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>